Programs can be client dependencies
