+++ ci/token-swap.sh
@@ -7,7 +7,7 @@ cd "$(dirname "$0")/.."
 ./do.sh update
 ./do.sh build token
 ./do.sh fmt token-swap --all -- --check
-./do.sh build-native token-swap -D warnings
+./do.sh build-lib token-swap -D warnings
 ./do.sh build token-swap
 ./do.sh clippy token-swap -- --deny=warnings
 ./do.sh doc token-swap
