+++ memo/Cargo.toml
@@ -12,9 +12,11 @@ edition = "2018"
 
 [features]
 skip-no-mangle = ["solana-sdk/skip-no-mangle"]
+program = ["solana-sdk/program"]
+default = ["solana-sdk/default"]
 
 [dependencies]
-solana-sdk = { version = "=1.2.13", default-features = false, features=["program"] }
+solana-sdk = { version = "=1.2.13", default-features = false, optional=true }
 
 [lib]
 name = "spl_memo"
