+++ token-swap/Cargo.toml
@@ -9,16 +9,19 @@ authors = ["Solana Maintainers <maintainers@solana.foundation>"]
 repository = "https://github.com/solana-labs/solana-program-library"
 license = "Apache-2.0"
 edition = "2018"
+exclude = ["js/**"]
 
 [features]
 skip-no-mangle = ["solana-sdk/skip-no-mangle"]
+program = ["solana-sdk/program", "spl-token/program"]
+default = ["solana-sdk/default", "spl-token/default"]
 
 [dependencies]
 num-derive = "0.2"
 num-traits = "0.2"
 remove_dir_all = "=0.5.0"
-solana-sdk = { version = "=1.2.13", default-features = false, features=["program"] }
-spl-token = { path = "../token" }
+solana-sdk = { version = "=1.2.13", default-features = false, optional = true }
+spl-token = { path = "../token", default-features = false, optional = true }
 thiserror = "1.0"
 
 [dev-dependencies]
