+++ token/Cargo.toml
@@ -13,12 +13,14 @@ exclude = ["js/**"]
 
 [features]
 skip-no-mangle = ["solana-sdk/skip-no-mangle"]
+program = ["solana-sdk/program"]
+default = ["solana-sdk/default"]
 
 [dependencies]
 num-derive = "0.2"
 num-traits = "0.2"
 remove_dir_all = "=0.5.0"
-solana-sdk = { version = "=1.2.13", default-features = false, features=["program"] }
+solana-sdk = { version = "=1.2.13", default-features = false, optional = true }
 thiserror = "1.0"
 
 [dev-dependencies]
